Rename classes and arguments of methods
