+++ docs/source/correctors.rst
@@ -1,0 +1,11 @@
+==============
+WCS Correctors
+==============
+
+.. moduleauthor:: Mihai Cara <help@stsci.edu>
+
+.. currentmodule:: tweakwcs.correctors
+
+.. automodule:: tweakwcs.correctors
+   :members:
+   :undoc-members:
