--- docs/source/tpwcs.rst
@@ -1,11 +1,0 @@
-=====
-tpwcs
-=====
-
-.. moduleauthor:: Mihai Cara <help@stsci.edu>
-
-.. currentmodule:: tweakwcs.tpwcs
-
-.. automodule:: tweakwcs.tpwcs
-   :members:
-   :undoc-members:
